Refactored and separated the DLLEncryptorLoader from DLLEncryptor impl
